feat: Interface 도입
