+++ src/test/java/repository/WiseSayingRepositoryTest.java
@@ -1,0 +1,5 @@
+package repository;
+
+class WiseSayingRepositoryTest {
+
+}
